Handle an edge case in the desc set allocation
There might be a situation in which the number of at least one type of descriptor set exceeds what is left of that des type in the pool, even though we can technically allocate many more sets from the pool if possible. In this situation we will get VK_ERROR_OUT_OF_POOL_MEMORY. So we need to handle this case by allocating another pool and advancing the relevant handles.